# core-common-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/distributors/{distributorId}/contracts/{contractId}:
    get:
      tags:
        - Distributor Contract
      summary: Get contract by ID
      description: Retrieve a specific contract associated with a distributor by its unique identifier
      operationId: getById_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Contract found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorContractDTO'
        '404':
          description: Contract or distributor not found
    put:
      tags:
        - Distributor Contract
      summary: Update contract
      description: Update an existing contract associated with a distributor
      operationId: update_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorContractDTO'
        required: true
      responses:
        '200':
          description: Contract successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorContractDTO'
        '400':
          description: Invalid contract data
        '404':
          description: Contract or distributor not found
    delete:
      tags:
        - Distributor Contract
      summary: Delete contract
      description: Delete a contract associated with a distributor from the system
      operationId: delete_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: contractId
          in: path
          description: Unique identifier of the contract
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Contract successfully deleted
        '404':
          description: Contract or distributor not found
...
  /api/v1/distributors/{distributorId}:
    get:
      tags:
        - Distributor
      summary: Get distributor by ID
      description: Returns a distributor based on its ID
      operationId: getDistributorById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor
      summary: Update distributor
      description: Updates an existing distributor with the provided information
      operationId: updateDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorDTO'
        required: true
      responses:
        '200':
          description: Distributor successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '400':
          description: Invalid distributor data provided
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor
      summary: Delete distributor
      description: Deletes a distributor based on its ID
      operationId: deleteDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Distributor successfully deleted
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/contracts:
    post:
      tags:
        - Distributor Contract
      summary: Create contract for a distributor
      description: Create a new contract associated with a specific distributor
      operationId: create_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorContractDTO'
        required: true
      responses:
        '201':
          description: Contract successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DistributorContractDTO'
        '400':
          description: Invalid contract data
        '404':
          description: Distributor not found
  /api/v1/distributors/{distributorId}/contracts/filter:
    post:
      tags:
        - Distributor Contract
      summary: Filter contracts for a distributor
      description: Retrieve a paginated list of contracts associated with a specific distributor based on filtering criteria
      operationId: filter_1
      parameters:
        - name: distributorId
          in: path
          description: Unique identifier of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorContractDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered contracts
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Distributor not found
...
  /api/v1/distributors/filter:
    post:
      tags:
        - Distributor
      summary: Filter distributors
      description: Returns a paginated list of distributors based on filter criteria
      operationId: filterDistributors
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributors
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
...
  /api/v1/lending-contracts/distributor/{distributorId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contracts by distributor
      description: Returns all lending contracts associated with a specific distributor
      operationId: getLendingContractsByDistributorId
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
  /api/v1/lending-contracts/contract/{contractId}:
    get:
      tags:
        - Lending Contract
      summary: Get lending contract by contract ID
      description: Returns a lending contract based on its contract ID
      operationId: getLendingContractByContractId
      parameters:
        - name: contractId
          in: path
          description: Contract ID of the lending contract to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved lending contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingContractDTO'
        '404':
          description: Leasing contract not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    LendingContractDTO:
      required:
        - contractId
        - distributorId
        - downPayment
        - endDate
        - lendingConfigurationId
        - monthlyPayment
        - partyId
        - productId
        - startDate
        - totalAmount
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        contractId:
          type: string
          format: uuid
        partyId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        lendingConfigurationId:
          type: string
          format: uuid
        originatingAgentId:
          type: string
          format: uuid
        agencyId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        monthlyPayment:
          type: number
        downPayment:
          type: number
        totalAmount:
          type: number
        status:
          pattern: DRAFT|PENDING|APPROVED|ACTIVE|COMPLETED|CANCELLED|TERMINATED
          type: string
        approvalDate:
          type: string
          format: date-time
        approvedBy:
          type: string
          format: uuid
        termsConditions:
          maxLength: 5000
          minLength: 0
          type: string
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          format: uuid
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: string
          format: uuid
          readOnly: true
    DistributorDTO:
      required:
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalCode:
          maxLength: 255
          minLength: 0
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        displayName:
          maxLength: 255
          minLength: 0
          type: string
        taxId:
          type: string
        registrationNumber:
          maxLength: 100
          minLength: 0
          type: string
        websiteUrl:
          maxLength: 255
          minLength: 0
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        supportEmail:
          type: string
        addressLine:
          maxLength: 255
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        state:
          maxLength: 100
          minLength: 0
          type: string
        countryId:
          type: string
          format: uuid
        isActive:
          type: boolean
        isTestDistributor:
          type: boolean
        timeZone:
          maxLength: 50
          minLength: 0
          type: string
        defaultLocale:
          maxLength: 10
          minLength: 0
          type: string
        onboardedAt:
          type: string
          format: date-time
        terminatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    DistributorContractDTO:
      required:
        - contractNumber
        - contractType
        - distributorId
        - endDate
        - startDate
        - title
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        distributorId:
          type: string
          format: uuid
        contractNumber:
          type: string
        contractType:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        renewalDate:
          type: string
          format: date
        noticePeriodDays:
          type: integer
          format: int32
        autoRenewal:
          type: boolean
        contractValue:
          type: number
        currencyCode:
          type: string
        paymentTerms:
          type: string
        specialTerms:
          type: string
        financialConditions:
          type: string
        productConditions:
          type: string
        serviceLevelAgreements:
          type: string
        metadata:
          type: string
        signedDate:
          type: string
          format: date
        signedBy:
          type: string
          format: uuid
        approvedDate:
          type: string
          format: date
        approvedBy:
          type: string
          format: uuid
        terminatedDate:
          type: string
          format: date
        terminatedBy:
          type: string
          format: uuid
        terminationReason:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          format: uuid
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: string
          format: uuid
          readOnly: true
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestDistributorContractDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorContractDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for contracts
...
    FilterRequestDistributorDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...